improved buffered io status, minor efficiencies in string usage and buffer reallocation

+++ src/fileio.h
@@ -8,7 +8,7 @@
 
 #include "bufferedio.h"
 
-size_t fio_read_all(buffer_t *buf, const char *path);
+int fio_read_all(buffer_t *buf, const char *path);
 
 size_t fio_write_all(const char *path, buffer_t *buf);
